Avoid re-allocating a pool for an Activity/Service that has been destroyed
Summary:
When a ViewCompatComponent gets measured, it acquires a MountContent to be able to measure it. If this happens after the Context has been destroyed (this can happen with an IdleExecutor for example), it might re-register the Activity in ComponentsPools and leak it.
This diff makes sure that a content is only acquired during measurement if it's safe to do so (the activity has not been destroyed). Unfortunately we can't do much for a Service as there is no way to check if a Service has been destroyed.

Reviewed By: marco-cova

Differential Revision: D5102276

fbshipit-source-id: 4844b86d363bcd7972cacbddd5dcaff2eda55f89