Handle long clicks properly
Summary: Right now, if you have a text drawable with a clickable span, then when you long click on it you don't get the long click event, since the event gets taken by the clickable span. Instead we need to do something similar to https://fburl.com/bp95edod and let long clicks be listened for, and if they take place then we should consider the touch event to be over.

Reviewed By: pasqualeanatriello

Differential Revision: D5086925

fbshipit-source-id: 52b2a082685c0cfcc2caed2ec26a0224f79714d2