the root host should always be marked as mountable
Summary:
In some cases we were ending with the root host marked as mounted but not mountable (its bounds were not intersecting the visible bounds). In this scenario we were trying to "unmount" the root host (that cannot really be unmounted) and we were skipping setting all the view attributes.
In any following incrementalMount step the root host would still appear as mounted but since the MountState is not dirty anymore at that point we would never update the view attributes anymore.

Reviewed By: IanChilds

Differential Revision: D5112274

fbshipit-source-id: c3a80c66af63bad509d710c5be07d7819ab2c249