Infra for exposing Diff of @Prop/@State params in lifecycle methods
Summary:
Adds infra support for receiving the previous version of the props/state a component rendered with in a lifecycle method.

The reason for this is so that we know when to trigger more complex animations (e.g. sequences and staggers, animations with initial velocity, etc) instead of always triggering and trying to magically reconcile the animations. This is the same strategy that ComponentKit and React take. This data can also be used to trigger events exactly once if we add a 'componentWillUpdate callback'.

Annotation Processor support is coming in a later diff.

Overview:
- If a component uses an Previous prop/state (I call this PreviousRenderInfo) we will generate a datastructure that holds exactly those props and state params that we need a previous version of.
- When calculating layout state, we record all the components that need previous render state.
- Just before mount, we walk those components and populate them with the last values they were mounted with. We use the same global keys StateHandler uses.
- Just after mount, we walk those same components and record the necessary props and state.
- If a component uses previous props and state, we defer calling the animation callback `onCreateTransition` until mount time (e.g. after we'll have previous mount data)

The reason we need to defer lifecycle methods that use previous render info until mount is because that's the only time we're guaranteed that the currently mounted state is the exact previous state.

Reviewed By: pasqualeanatriello

Differential Revision: D5095059

fbshipit-source-id: f6949ba08d52699eba9513012142931d20d19253