[WIP] Upgrade to Gradle Plugin 3.0
This is a work-in-progress for upgrading to the new gradle plugin. It
significantly improves build times as it no longer implicitly propagates
`compile` dependencies and doesn't run all annotation processors it can
find on the class path.

It should also allow us to get rid of the `RELEASE_MODE` hack as it
properly supports variant propagation.

Sadly, the tests are currently failing due to this error:

<details>
```
java.lang.IllegalArgumentException: The key must be an application-specific resource id.
	at android.view.View.setTag(View.java:14978)
	at com.facebook.litho.ComponentHost.setTag(ComponentHost.java:379)
```
</details>

My suspicion is that this is because of Robolectric and not something we
do, but I'd be happy to be proven wrong about this.

I'll leave this around here for comments and if someone has a bunch on
how to fix the resource lookup, please let me know!