Document alternative to local ruby installation
Summary:
I hope that will lower the barrier to entry/contribution somewhat.

FB-Only:
This is still not like clicking a button on a Dex, but setting up the Ruby
dependencies is a **real** pain in the face and this is now just one command.

This is in response to the comments on https://fb.facebook.com/notes/lucas-rocha/lets-talk-about-documentation/1818142048497498/

Reviewed By: IanChilds

Differential Revision: D5128866

fbshipit-source-id: e62753b93ce770ca90cb02800e3f43c375f9db18