Keep track of local transient state to decide when to disable incremental mount
Summary:
checking ViewCompat.hasTransientState will also track transient state for children of the LithoView. We are only really interested in knowing if transient state was set directly on the LithoView.
Before this diff if transient state was set on an inner child of a LithoView we would stop doing incremental mount without fully mounting the LithoView.

Reviewed By: IanChilds

Differential Revision: D5144937

fbshipit-source-id: a950ff4fe368fee1b2444e2c7939a4a337b2cc04