Fix crash when ComponentHost is trying to unregister touch expansion that hasn't been registered
Summary: There was a crash happening with items that have touch expansion registered. When mounting or unmounting LayoutOutputs we correctly register/unregister touchexpansion from their ViewNodeInfo. However, when we are updating existing MountOutput (that has moved to new position) we did not do anything about touchexpansion, therefore the touchexpansion from LayoutOutput was not registered to host's touchexpansiondelegate. And when trying to unregister such item, the app was crashing because touchexpansiondelegate did not have corresponding item. This diff fixes this issue by registering touch expansion if needed when updating MountOutput.

Reviewed By: marco-cova

Differential Revision: D5149043

fbshipit-source-id: f77aa611b9300e4681afb46e925cb216973a2bf4