Terminate displaylist creation in `DisplayListPrefetcher` if the activity `LayoutState` bound to has finished
Summary: There are cases when `LayoutState` was enqueued for displaylist calculation when app was in foreground, and `Runnable` that creates displaylist from it was actually executed when app went background. Imagine scrolling feed and exiting the app in the middle. When such case occurs GLContext might have been killed already by the time we are trying to create a displaylist, which will crash the app. This diff checks whether the activity that `LayoutState` is bound to is still alive before creating displaylist.

Reviewed By: pasqualeanatriello

Differential Revision: D5173127

fbshipit-source-id: a7d815bdd6c1361a95d7502191a82dfc2e6a670f