Support ResType on PropDefault annotation
Summary:
Initial work on adding support for `ResType` on `PropDefault` ( #81 )

I went with the solution of having the `Builder` initializing the Props that have a default value marked as ResType, instead of passing down the `ComponentContext` which I tried to implement but it was introducing a quite huge change.

In order to reduce the review scope I have only added generation for `resolveDimenSizeRes()` and will add the rest before closing the PR, in order to have a naming and feature compliance agreement after the initial review. Also, I believe that after resolving the resource on a marked `PropDefault` we should generate resource helper methods like we do for `Prop`.

Let me know what you think!

/cc: emilsjolander

P.S.: There is a big diff on the `ComponentImplGenerator` due to different Code Style, so If you could share the one you are using so that we could be aligned and reduce the diffs, would be great.
Closes https://github.com/facebook/litho/pull/174

Reviewed By: passy

Differential Revision: D5171855

Pulled By: IanChilds

fbshipit-source-id: 1d52389ce95b4694b6b8a1b8d95f50d334e9b2a3