Disable predictive animations for child Horizontal RecyclerViews when using binder with LinearLayoutInfo
Summary: There is a rare case when RecyclerView's internal state becomes inconsistent with adapter's state and app crashes. This happens only when one RecyclerView is used as a child of another one and when predictive animation is enabled. This issue is known and it's a bug in RecyclerView that hasn't been yet fixed properly (https://issuetracker.google.com/issues/37007605). This diff turns off predictive animation for Horizontal RecyclerViews which are used as an inner RecyclerView for most of our usecases.

Reviewed By: pasqualeanatriello

Differential Revision: D5200629

fbshipit-source-id: 5657e71fcf508b033214bcba61af336c173519b9