Add annotation processor support for @State Diff<T> stateParam/@Prop Diff<T> propParam parameters to lifecycle methods
Summary:
Depends on D5095059

See D5095059 for more info on why we want this. Diff params are able to give you both the current and previous values. This annotation processor support generates code that implements the methods added in D5095059 so that components that need Diff params can use them.

Reviewed By: IanChilds

Differential Revision: D5114655

fbshipit-source-id: 05fe7945e1d5cab559d6b0a4b4a0c06dc45f2f70