Add support for text touch offset change event in TextSpec
Summary: Here is the use case. An app using `Text` component wants to know the offset of the text under the touch point. While it would be possible to get (x,y) coordinates by adding `touchHandler` to `TextSpec`, it's hard to get offset of the text located on that (x,y) position because we don't expose mounted item (`TextDrawable`) outside `TextSpec`. Declarative way of achieving that would be through exposing events which is what this diff does.

Reviewed By: marco-cova

Differential Revision: D5236271

fbshipit-source-id: abdb72c9c551069793dcadebfac71ce49a0c1abb