Fix crash with getChildDrawingOrder
Summary: `getChildDrawingOrder` depends upon `mViewMountItems`, `mDisappearingItems` and `mScrapHosts`. Therefore we should always set `mIsChildDrawingOrderDirty` to true when one of these changes. Right now, in this code path we were setting it to true at the start of `unmountView`, but on new versions of Android (7 and above I think) the removeView code calls into accessibility code which causes a call to getChildDrawingOrder. That resets the flag and so it is not updated after the view is removed from `mViewMountItems`, which is wrong.

Reviewed By: marco-cova

Differential Revision: D5245663

fbshipit-source-id: f77a2c6d1247cf9ca106cff7377d354b09327bdc