Move creating the EditTextSpec eventHandler in bind
Summary:
The EditTextSpec eventHandler was being created in onMount. This was causing issues with rendering EditText components that received textChangedEventHandlers that depended on props, because we never reset the handler that we used for dispatching the TextChangedEvents when props would change. The event handler would receive stale props because of this.
Context: https://fburl.com/etqp4xae

Reviewed By: IanChilds

Differential Revision: D5256094

fbshipit-source-id: 6c1f10ecdd603c2b845ad468ea7e340137c670e3