Detach LithoView of componentTree before using the tree with sticky header
Summary:
Sticky header in sample app had a bug with RecyclerView v25 when scrolling up and then down made RecyclerView's first child's height 0 which resulted in not being able to scroll to the top of the RecyclerView's first item. The video of the bug repro can be found in the github issues: https://github.com/facebook/litho/issues/101.
What was happening is that the view that went out of screen was not detached by the time we used its componentTree for overlay sticky header. When we use componentTree whose view is still attached to window to set to overlay sticky header view, it will clear previous view's componentTree, and next time it comes to screen it's height will be 0. Detaching the view when initializing sticky header solves this issue.

Reviewed By: pasqualeanatriello

Differential Revision: D5245439

fbshipit-source-id: 2b77903f3fa0ad635e7a4cb13045954aca15d2a4