Fix NPE when trying to create displaylist with null context
Summary: There is a small chance that when we are creating displaylist the `LayoutState` might have been already released. In fact we are checking that inside `DisplayListPrefetcher.run()` by calling `currentLayoutState.isActivityValid()`, however, the `LayoutState` can be accessed from background thread and could be released before we reached `createDisplayList()` method in the prefetcher.

Reviewed By: marco-cova

Differential Revision: D5274062

fbshipit-source-id: dfc4e7fb8209f10941f8bf011f4da654f2fc55d8