Upgrade to gradle plugin version 3
Summary:
Yes, this is still an alpha. But not only does it allow to use the new features
of Android Studio right now, but we also get a significant build time speedup.

We now consistently build in less than 20 minutes, often less than 18 minutes;
down from 28 to 30 minutes. That's nothing to sneeze at.

The plugin also addresses an old issue we had with annotation processors running
implicitly and causing tests to fail we (well, Ian) then had to rewrite.

We managed to report some bugs upstrem and had the Android team fix them, so
there seems to be a lot of value in adopting this early. Especially as it's not
sitting in the critical path for our team.

Reviewed By: IanChilds

Differential Revision: D5274572

fbshipit-source-id: 0aaff425ce0cf6448c4a339c28534076bcc708b2