Skip LithoViewTestHelperTests in release mode
Summary:
I forgot that we run gradle tests in both debug and release mode. For
LithoViewTestHelperTest this means that we won't have access to debug nodes in
the release tests, meaning they will fail - which is now happening in Circle CI.

This skips the tests in release mode, as they're pointless in that case.

Reviewed By: pasqualeanatriello

Differential Revision: D5329524

fbshipit-source-id: 00c6a74bbaa8cfb205eeb97310770b8bb22e4012