Open-source processor integration tests
Summary:
Ooooph, that wasn't nearly as straightforward as I had hoped.
There's a bunch of magic around those tests as they not only need to be compiled
with Java8 to work with Truth, but also cannot run as "robolectric_test" in the
Buck sense but still *require access to Android APIs*. This now runs as regular
`java_test`, but pulls in Robolectric as it exports the Android API JAR.

I also had to change the resource loading as gradle requires a very particular
place and Buck on the other hand does not work with files that aren't below its
current working directory. In addition to that, I had to include a hack for
gradle to prevent it from *compiling* the Java resources we have. Renaming the
files to e.g. `.java_` won't work either as Truth/Compile-Testing won't
recognize them as Java sources in that case.

Reviewed By: IanChilds

Differential Revision: D5328871

fbshipit-source-id: e5804af476e7a835a9d99b60b7a5e1637ffae796