Introduce DisplayListContainer shared between LayoutOutput and MountItem
Summary:
This diff creates a data structure (`DisplayListContainer`) that is shared between `LayoutOutput` and `MountState` and contains `DisplayList` object. We will always create `DisplayListContainer` when initializing `LayoutOutput` if the component supports displaylists.

That is necessary for further work where we would like to cache the displaylist of just drawn component and draw displaylist directly next time. When the component is drawn, the displaylist of that component is saved in `DisplayListContainer` and even if we remount component we will still have valid displaylist of that component.

Reviewed By: pasqualeanatriello

Differential Revision: D5312779

fbshipit-source-id: 6e9d44da6a04fd6cfd20acea9ee672578399df39