Do not fire textOffsetOnTouch event when the motionevent is outside drawable bounds
Summary: We were firing textOffsetOnTouch event even when the motion event was outside the bounds. If component has three Text items and when we press on one of them, the event was delivered to all of them, because their host was same and all touch events are delegated through component's host. This diff adds bounds check.

Reviewed By: astreet

Differential Revision: D5365634

fbshipit-source-id: b13e200eb7aee6f6283cdaa396c37589a0813190