Posting sync state updates on caller thread
Summary:
Right now, triggering state updates from visibility event handlers can cause a new layout calculation to be started before the one that triggered the updates finished, so we could be trying to finish mount on an invalid LayoutState.
This is an attempt to post sync state updates on the calling thread, that way we first finish the layout calculation and then attempt to apply any pending state updates.

Reviewed By: pasqualeanatriello

Differential Revision: D5010787

fbshipit-source-id: fe2009580ba9910e60ffd82d1c12dfb63d40bf5b