Don't try to incrementally mount LithoView if the ComponentTree doesn't have a valid MainThreadLayoutState
Summary:
Alternative/addition to D5369274

Makes it so we just bail if we try to incrementally mount a ComponentTree that doesn't have a main thread layout state. The alternative would be to go ahead and create it using the LithoView's current dimensions, but there's a good chance we'd re-compute it again when the view actually lays out.

Reviewed By: pasqualeanatriello

Differential Revision: D5369314

fbshipit-source-id: bb372cf2eb1f1a59c70f59323d3626f8a1da7959