Add mAnimationLockedIndices data structure to keep content from incremental-unmounting during animation
Summary:
Animation-locked indices are the indices of mount items we know are currently animating, or are the parent or child of an animating mount item. Any locked index will not be unmounted due to visibility reasons (e.g. incremental mount). We do this because the view may actually still be on screen due to an X/Y position animation.

We ref-count the locking because an index may be locked because of multiple different animations.

Reviewed By: marco-cova

Differential Revision: D5346890

fbshipit-source-id: e3d81f4ef52d217cbf3606c4fd00f9a4551fceab