Remove onNewTransitionContext and onPre/PostMount from TransitionManager
Summary:
Depends on D5346890

To make the diffs easier to read, I'm splitting up removing the old code for recording diffs and adding the new code for recording diffs (phabricator was trying to half-reuse a bunch of lines which made it really hard to tell what was being added).

These removed methods will be replaced in the next diff.

Reviewed By: marco-cova

Differential Revision: D5347198

fbshipit-source-id: f6a8b9a631076c5d4d5f302a3e15268ee6b72611