Use a single listener to communicate the ends of animations back to MountState
Summary:
Depends on D5349209

The previous listener worked with mount content while we want to work with LayoutOutputs and transition keys. The basic premise is the same, we'll just have a single listener in MountState that gets called whenever an animation ends. MountState can then do any cleanup it needs to there.

Reviewed By: marco-cova

Differential Revision: D5349410

fbshipit-source-id: ed6beb3424f30fab8d9bfd6c16423781af835c24