Remove isIncrementalMountOnOffsetOrTranslationChangeEnabled flag
Summary:
This has always defaulted to on, and we just had the flag there so that we could turn it off if we saw weird things. Thankfully we didn't, so we can get rid of it.

ListView doesn't make the same calls that RecyclerView does, so ListView will have to continue to use the scroll listener. Essentially, anywhere that calls `IncrementalUtils.performIncrementalMount(..., true)` remains, but the others can be deleted as they weren't doing anything anyway.

Reviewed By: pasqualeanatriello

Differential Revision: D5355729

fbshipit-source-id: 52f977a8734b0b2c0574ff0e29372b9c969e85a3