Add e2e screenshot test for transition animations
Summary:
Adds an e2e screenshot test for transition animations.

The animation we use is from the transitions demo in the sample app. The animation is driven by a MockTimingSource which allows the test to manually step through the animation via `stepAnimationFrame`, taking a screenshot at key points. This first test tests scrolling an animating view off the screen during animation (e.g. testing the interaction with incremental mount).

Reviewed By: passy

Differential Revision: D5443173

fbshipit-source-id: ef54beb05ee42e3f968a92c17cefa546c2e836b1