Make Typeface threadsafe
Summary:
Typeface isn't threadsafe, so if you call Typeface.create off the main thread, you can get crashes with accessing SparseArrays.

This is an attempt to fix that by replacing the implementation of sTypefaceCache with something that is synchronous. It also makes sure that any SparseArrays that are returned from the LongSparseArray are synchronous, which means that there shouldn't be threading issues on them.

Reviewed By: sjkirby, pasqualeanatriello

Differential Revision: D5347724

fbshipit-source-id: ff1ef9aac2ec7e648c4ae9fbd5328ab5808eaa14