Set NO_POSITION as default first visible position in RecyclerBinder
Summary: We were setting default value of `mCurrentFirstVisiblePosition` as 0 and that had unintended side effect. When we have no children and `RecyclerView` is mounted, `RecyclerBinder` calls `scrollToPosition()` and passes `mCurrentFirstVisiblePosition` which is `0`, even though `RecyclerView` has no children yet. If `RecyclerView`'s `.stackFromEnd()` property is set to `true`, current implementation was causing `RecyclerView` not to stack from end when it is shown first time, as it internally was keeping `pendingScrollToPosition` and was ignoring `stackFromEnd` property if that value was non-negative. Current implementation did not cause any issue so far yet because all the surfaces where we used `RecyclerBinder` assumed stacking from top and having currentFirstVisiblePosition as 0 did not change much. This diff sets the default value of `mCurrentFirstVisiblePosition` to NO_POSITION (-1) so that when `RecyclerView` is mounted first time we don't call `scrollToPosition()`, which works for both cases when `RecyclerView` is stacked from top and bottom.

Reviewed By: mihaelao

Differential Revision: D5453147

fbshipit-source-id: 8107c94270fb3a473332bd66cfb354313d402285