Add two event tests to the sample app
Summary:
I added two tests that trigger the event handler in the item component which
show two different ways of achieving the same goal. I want to pick this up later
and find a nicer way of doing this, especially one where we don't have to rely
on Robolectric to set up Contexts and capture started activities.

Reviewed By: IanChilds

Differential Revision: D5488469

fbshipit-source-id: 2a21f2d5d7913c060a0c3506ca26370d89f77cc0