Use a new state handler instance to calculate layout
Summary:
There's a race condition with multiple news feed component trees that share the same persistent state StateHandler instance where one could be releasing the state handler while the other is accessing it in the LayoutState. This creates a new state handler instance before passing it to the component context and should mitigate that issue.
jingbo If this doesn't bring down the crashes I'll send another diff to not null out the data structures held by the state handler.

Reviewed By: muraziz

Differential Revision: D5488773

fbshipit-source-id: 24da11dee271cd778d141ddbaafebea082d66177