Small fix for requestFocusWithOffset
Summary:
I'm using `requestFocusWithOffset` for a hscroll which snaps its items to the center (not the start or end), and also may have an initial position other than 0 (https://fburl.com/enahnj9k). This mostly worked as expected but while stress-testing it I noticed that in some cases the offset wasn't respected and the focused item was aligned to the start (left) of the hscroll instead of being centered.

After some debugging I found a code path that leads to this - if `RecylerBinder.requestFocusWithOffset` gets called before `mount`, then it stores the requested focus in a member variable and applies it in `mount`. There's also a member variable for the requested offset but it wasn't being updated, so the proposed fix is to do that.

Differential Revision: D5481106

fbshipit-source-id: 4f3004361c18ed861a6abb324aaa36beabf2c3c9