Remove event handlers taking Component<?>
Summary:
This removes the generators for event handlers that take a Component (as opposed to a ComponentContext). Because of the untyped nature of this, this causes a runtime class cast exception as the first `param` of the handler would always be expected to be a ComponentContext, which this code path doesn't provide.

Ian also did some digging and found out that they were introduced with D2850655 but without any clear indication as to why they were added.

Reviewed By: emilsjolander

Differential Revision: D5479932

fbshipit-source-id: cd27ac6beb5ca4c1a84bb271ee02ec38db193f67