Varag Props Code Gen
Summary:
Props that are annotated as varags will now have a different prop assignment in the builder.

When the prop of type `List` is annotated with as varags, we were assigning it straight to the ComponentImpl variables. This was confusing because the code below:
```
Recycler.onScrollListener(listener)
       .onScrollListeners(listOfListeners) // will implicitly remove the assignment above
```
would remove the first single listener added to the list.

We will now handle the prop differently by adding to an existing list if there is one, otherwise we will assign to the ComponentImpl's variable directly.

Reviewed By: muraziz

Differential Revision: D5505322

fbshipit-source-id: 5497655d989aac321f3162e4922ecc8141b178c8