Don't store appearFrom/disappearTo on animation bindings
Summary:
Continuing to break our dependence on the original Animated APIs: the plan is to store appearFrom/disappearTo values directly on the Transition objects themselves (see D5500828). This means that the AnimationBinding objects will only be responsible for running the animation and will not have knowledge of whether a appear/disappear value had to be filled.

I'll be reimplementing the appear/disappear logic in an upcoming diff.

Reviewed By: muraziz

Differential Revision: D5500962

fbshipit-source-id: 29d4e396cffb9277f2c0dc1bbe155ba88910fbbc