Replace ComponentProperty with PropertyAnimation
Summary: Replaces ComponentProperty with PropertyAnimation. They basically track the same thing, but ComponentProperty had the requirement that it has a reference to an AnimatedComponent which wasn't a good requirement. PropertyAnimation is just a POJO holding transitionKey, property, and the target value. We give a target value now so that the AnimationBindings don't have to be aware of appearFrom/disappearTo values.

Reviewed By: muraziz

Differential Revision: D5501377

fbshipit-source-id: bc8a5fe5d2fa43f067c3261a03ae6ed32f02aaef