Refactor and simplify TransitionsManager to not depend on Animated API
Summary:
This is the main diff to review.

This is the end result of all the smaller diffs in this stack: we can now create and run animations more directly from TransitionManager. This restructuring makes things more clear. I've added comments in code to try to help explain the new architecture but I'll reiterate it here:

Transition keys are 1-1 mapped to AnimationState
An AnimationState has many PropertyStates (one for each property)
A PropertyState can have up to one animation.

An animation state keeps track of the current mount content object, as well as the state of all animating properties (PropertyStates). A PropertyState keeps track of a AnimatedPropertyNode (which has the current value of that property in the animation), and a single animation and end value. This means that at any point in time, we're able to tell what animation is animating what property(s).

This gives us the ability to check, for example, whether a property is already animating to the target end value and choose not to create an animation (this will be necessary for some of the 'targeting' API changes that are planned).

The other main change is the way that appearFrom/disappearTo values are applied: they are read off the Transition object and not given to the animation itself anymore. For appearFrom values, we just set the appearFrom value in restoreInitialStates. For disappearTo, we just use that value as the target.

Reviewed By: marco-cova

Differential Revision: D5502454

fbshipit-source-id: 8952d123082178c7ea7a9faadf9fb65098438478