Add targeting API for animations
Summary:
Adds a targeting API for `OnCreateTransition`. It allows single, sets, and all components/properties to be targeted by a single transition. This API is backwards compatible with existing usage (single component, single property).

Examples of this API:

Animate 'all' properties on all components with transition keys:
```
Transition.create(Transition.allKeys())
    .animate(Transition.allProperties())
```

Animate the X property on all components with transition keys:
```
Transition.create(Transition.allKeys())
    .animate(AnimatedProperties.X)
```

Animate all properties on the component with transition key 'red':
```
Transition.create('red')
    .animate(Transition.allProperties())
```

Animate X and Y on the components with transition keys 'red' and 'blue':
```
Transition.create('red', 'blue')
    .animate(AnimatedProperties.X, AnimatedProperties.Y)
```

Note: this does not affect appearFrom/disappearTo yet, e.g. saying to animate all properties does not add appear/disappear animations at this time. This might be something we want to consider in this api for the future but doesn't need to be addressed in this diff.

Reviewed By: marco-cova

Differential Revision: D5530528

fbshipit-source-id: 8bfab879120e19d43f4b3fbee6baa4807db53207