Log soft error for duplicate key
Summary:
before landing D5381379, we need to make sure that all sibling components of the same type have unique keys even if they are stateless. This logs a soft error when that happens and tries to build a useful stack trace to make it easier to find and fix the root of the duplication issue.

The plan is to use the soft errors to detect the duplications, then when those are mitigated land D5381379.

Reviewed By: emilsjolander

Differential Revision: D5536169

fbshipit-source-id: b4be578a7fbc943bde063e4c6c83fa908ecbe23b