Unmount child LithoViews of the unmounted views
Summary:
When a mounted view is unmounted, we recycle them into the pools and we can reuse next time we need such content. If such view contains other LithoViews as children those children are not unmounted when their parent is unmounted, hence we don't get the child items into the pool, but rather we'll throw them away.
Consider scenario where RecyclerView contains another RecyclerView (HScroll) as a child. When inner RecyclerView is unmounted, its children are not unmounted.

This diff adds `HasLithoViewChildren` interface that mount objects (ViewGroups) can implement to provide child `LithoView`s which will be unmounted when their parent is unmounted.

Reviewed By: IanChilds

Differential Revision: D5537663

fbshipit-source-id: a5c1403f113d4d15e2a28f59530842f110d42e64