implement special getProps() for DiffSectionSpecModel to create diff props
Summary: DiffSectionSpecs is special in that all of it's props are implicitly typed: `Diff<?>`.  That means `Prop Diff<Object> k` creates a builder method `Builder k(Object kProp)`.  The current processor architecture treats `Prop Diff<Object> k` as a `DiffModel` instead of a `PropModel`, this diff works around this by implementing it's own getProps() that grabs all the diff modesl and strips away the `Diff<?>`.

Differential Revision: D5493757

fbshipit-source-id: 2a73d7cbb8f5517ea4bc721bf1a4b3b4cc61af45