Update onCreateTransition to return abstract Transition class
Summary:
This diff will be part of a stack. My goal is to make it so the API for animations supports all the following usages:

A single transition:
```
OnCreateTransition
static Transition onCreateTransition(ComponentContext c) {
  return Transition.create("myKey")
      .animate(Transition.allProperties());
}
```

A parallel set of transitions:
```
OnCreateTransition
static Transition onCreateTransition(ComponentContext c) {
  return Transition.parallel(
      Transition.create("first")
          .animate(Transition.allProperties()),
      Transition.create("second")
          .animate(Transition.allProperties()));
}
```

A sequence of transitions:
```
OnCreateTransition
static Transition onCreateTransition(ComponentContext c) {
  return Transition.sequence(
      Transition.create("first")
          .animate(Transition.allProperties()),
      Transition.create("second")
          .animate(Transition.allProperties()));
}
```

A sequence of parallel transitions (and vice-versa):
```
OnCreateTransition
static Transition onCreateTransition(ComponentContext c) {
  return Transition.sequence(
      Transition.parallel(
          Transition.create("first")
              .animate(AnimatedProperties.X),
          Transition.create("second")
              .animate(AnimatedProperties.X)),
      Transition.parallel(
          Transition.create("first")
              .animate(AnimatedProperties.Y),
          Transition.create("second")
              .animate(AnimatedProperties.Y)));
}
```

In order to accomplish this, I'm doing the following:
- Renaming `Transition` to `SingleTransition`
- Making `TransitionSet` and `SingleTransitionBuilder` extend `Transition` so that both single transitions and sets of transitions can be composed in a sequence or parallel set

For now, this diff assumes a `TransitionSet` of `SingleTransitions` is returned (which is always the case before this diff anyway). That assumption will be updated in the following diffs in this stack.

Reviewed By: marco-cova

Differential Revision: D5554388

fbshipit-source-id: 53fd7ec4944acbd7f4ef3268701cfe8a8930a038