Avoid NPE when asking for name of garbage collected node
Summary: If a race condiftion ocurred and a node was garbage collected before a rerender of the debug UI had ocurred then we would be accesing a null node to show the name of the component. To ensure this does not happen we instead save the class of the node on creation of the debug component.

Differential Revision: D5570898

fbshipit-source-id: bf9c072dbd97bdd374efc4cdf1896b8bdbbbb63a