Add support for nested sequences/parallel sets of animations
Summary:
This diff is part of the work needed to support sequence animations. As mentioned in D5554388, we want to support transitions of the form:

```
OnCreateTransition
static Transition onCreateTransition(ComponentContext c) {
  return Transition.sequence(
      Transition.parallel(
          Transition.create("first")
              .animate(AnimatedProperties.X),
          Transition.create("second")
              .animate(AnimatedProperties.X)),
      Transition.parallel(
          Transition.create("first")
              .animate(AnimatedProperties.Y),
          Transition.create("second")
              .animate(AnimatedProperties.Y)));
}
```

To that end, this diff adds support for resolving the nested structure of one of these Transition trees. It also makes `createSet` return a ParallelTransitionSet which is turned into a ParallelBinding animation (which is the dataflow support for parallel sets of animations). After this diff, I'll rename `createSet` to `parallel` and also introduce `sequence`.

Reviewed By: marco-cova

Differential Revision: D5557071

fbshipit-source-id: 4f19889b1ca6f3b8715f78e9ce07692ca7508f4f