Replace all usages of ComponentInfo.create().build() with ComponentInfo.createEmpty()
Summary: We want to enforce a rule not to create empty `ComponentInfo` using builders, but rather use a special method for such purpose. This will be needed in later diffs where `ComponentInfo` will be used to contain either component or viewbinder but they shouldn't be null both at the same time.

Reviewed By: IanChilds

Differential Revision: D5582822

fbshipit-source-id: f05ea81304e53e137cb25f6f92afd0b0241fd577