Fix interrupting sequence/stagger animations
Summary:
API support for sequences and staggers is landing, but there are issues if a sequence/stagger animation gets interrupted (i.e. if one of unstarted constituent animations becomes invalid because a new layout state is mounted).

To solve this, when an animation is about to run, we now verify it's still valid.

If it's not, we do the equivalent of immediately ending it.

This probably isn't the optimal reconciliation strategy, but it's good enough for now as it means that we always eventually show the right mount content.

Reviewed By: marco-cova

Differential Revision: D5574562

fbshipit-source-id: 04bc0ce2f091c6126f639ac52ade1b919984c4ac