Fix animations with incremental mount when child LithoViews are present
Summary:
When an animation takes place, we want to mount everything on the LithoView that is being animated. This is because animations won't call `performIncrementalMount` as the animated View moves on or off screen, so we might not mount some content that we need.

Right now, we try to force the whole LithoView to be mounted by passing a `null` rect. However, that causes issues because a `null` rect implies that we're not doing incremental mount and this is turn means that we don't bother mounting child views that have `canMountIncrementally` set to true.

Instead, let's do incremental mount but force the mount to take place on the whole LithoView. This will ensure that recursive mounting works too.

Reviewed By: pasqualeanatriello

Differential Revision: D5591947

fbshipit-source-id: 826f9ed631ed52ca5194c028e5627c5337abb505