Require debug mode to run SubComponent tests
Summary:
The new Sub Component testing API only works with
ComponentsConfiguration.IS_INTERNAL_BUILD. This also answers one of the
questions I left for myself in the code: In which circumstances can
getRootInstance() actually return null. It's when we don't have an internal
build.

Circle CI caught this as the `releaseDebugUnitTest`s failed.

Reviewed By: IanChilds

Differential Revision: D5592499

fbshipit-source-id: fede85c557c1ab39754341d48b29a73ec7814c33